Add 32-bit variant of TestU01 Rabbit battery
Nearly all tests defined in TestU01 are flexible enough to use 31-bit or even
lower resolution RNGs. However, the predefined batteries cannot be externally
configured with that resolution.

This commit adds a modified version of bbattery.c to the build which has a
31-bit variant of Rabbit. The original battery assumes you have 32-bit blocks
and so most of tests are configured for 32-bit resolution, causing many spurious
failures when using a 31-bit RNG.

As I am not the original author and do not have much experience with this kind
of math code, you should assume I made mistakes in the modification. It doesn't
crash but the math may be incorrect and should be reviewed.